+++ src/utilities/grid/Grid.root.tsx
@@ -1,0 +1,38 @@
+import styled, { css } from 'styled-components';
+
+import { Utils } from '../../shared';
+import { applyMargin, applyPadding } from '../spacing';
+
+import { applyGrid, applyGridItem } from './Grid.mixins';
+import { PApplyGrid, PApplyGridItem } from './Grid.props';
+
+const GridRoot = styled.div.withConfig({
+    shouldForwardProp: (property, validator) =>
+        Utils.blockProperty(property, [
+            'columnsTemplate',
+            'rowsTemplate',
+            'areasTemplate',
+            'gap',
+            'placeItems',
+            'placeContent',
+        ]) && validator(property),
+})(
+    ({ padding, margin, ...rest }: PApplyGrid) => css`
+        ${applyGrid(rest)};
+        ${padding && applyPadding(padding)};
+        ${margin && applyMargin(margin)};
+    `
+);
+
+const GridItemRoot = styled.div.withConfig({
+    shouldForwardProp: (property, validator) =>
+        Utils.blockProperty(property, ['columns', 'rows', 'area']) && validator(property),
+})(
+    ({ padding, margin, ...rest }: PApplyGridItem) => css`
+        ${applyGridItem(rest)};
+        ${padding && applyPadding(padding)};
+        ${margin && applyMargin(margin)};
+    `
+);
+
+export { GridRoot, GridItemRoot };
